feat: add deps for rete
