+++ package.json
@@ -50,11 +50,16 @@
     "react-toastify": "^10.0.4",
     "reactflow": "^11.9.2",
     "recharts": "^2.4.3",
+    "rete": "^2.0.3",
+    "rete-area-plugin": "^2.0.4",
+    "rete-connection-plugin": "^2.0.1",
+    "rete-react-plugin": "^2.0.5",
+    "rete-render-utils": "^2.0.2",
     "socket.io-client": "^4.6.1",
     "sphinx-bridge": "0.2.64",
     "stream-browserify": "^3.0.0",
     "string_decoder": "^1.3.0",
-    "styled-components": "^5.3.9",
+    "styled-components": "^6.1.9",
     "three": "^0.143.0",
     "three-forcegraph": "^1.39.4",
     "three-spritetext": "^1.6.5",
